The first rule of good C programming is to make sure that your variables, especially pointers, are initialised to a known state.
Fix a bug that meant bad memory seeped back into the memory manager ultimately causing it to crash. This was manifested in the situation when cmail was included into a larger program - such as a web application.

The fix causes the global script and the global closure for output to be set to NULL, so that when a default output closure is registered the system dies OR causes memory issues.